separate workload functions
